+++ aigilas/src/aigilas/skills/SkillInfo.java
@@ -8,13 +8,13 @@ import java.util.List;
 public class SkillInfo {
     public final String Name;
     public final StatType Stat;
-    public final float Cost;
-    public final float Magnitude;
+    public final int Cost;
+    public final int Magnitude;
     public final List<Elements> Elements;
     public final boolean Restrict;
     public final boolean OffCenter;
 
-    public SkillInfo(String name, StatType stat, float cost, float magnitude, List<Elements> elements, boolean bossOnly, boolean offCenter) {
+    public SkillInfo(String name, StatType stat, int cost, int magnitude, List<Elements> elements, boolean bossOnly, boolean offCenter) {
         Name = name;
         Stat = stat;
         Cost = cost;
